+++ internal/provider/data_source_iosxe_aaa_test.go
@@ -42,6 +42,7 @@ func TestAccDataSourceIosxeAAA(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_aaa.test", "group_server_radius.0.ip_radius_source_interface_loopback", "0"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_aaa.test", "group_server_tacacsplus.0.name", "tacacs-group"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_aaa.test", "group_server_tacacsplus.0.server_names.0.name", "tacacs_10.10.15.12"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_aaa.test", "group_server_tacacsplus.0.ip_tacacs_source_interface_loopback", "0"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -76,6 +77,7 @@ func testAccDataSourceIosxeAAAConfig() string {
 	config += `		server_names = [{` + "\n"
 	config += `			name = "tacacs_10.10.15.12"` + "\n"
 	config += `		}]` + "\n"
+	config += `		ip_tacacs_source_interface_loopback = 0` + "\n"
 	config += `	}]` + "\n"
 	config += `}` + "\n"